Include standard error for burden and VT
Note that the effect and standard error returned for VT are those
corresponding to the burden test over all minor allele frequency
thresholds that had the largest test statistic.

Recall that since VT computes the p-value from the MVN and covariance of
all test statistics over all burden tests, it will not correspond to
finding the p-value for just the largest test statistic T = effect / se.